requirements.txt
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
openai-whisper==20240918
pyannote.audio==3.1.1
pyannote.core
pyannote.metrics
pyannote.pipeline
speechbrain==1.0.0
jinja2==3.1.5
pydub
ffmpeg-python
soundfile
tqdm
